move: baseEntity 디렉토리 변경
