+++ src/modules/comment/entity/comment.entity.ts
@@ -1,5 +1,5 @@
 import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
-import { BaseEntity } from "../../../global/common/base.entitiy";
+import { BaseEntity } from "../../../global/common/entity/base.entitiy";
 import { Board } from "../../board/entity/board.entity";
 import { User } from "../../user/entity/user.entity";
